Refactor output file generation.
- Extract output file generation into separate function and file.

- Begin moving some business logic from runner to optimizer.
